Add: ダークモードを実装

+++ app/src/hooks/talk/mutate/index.ts
@@ -13,8 +13,24 @@ const postTalk = async (talk: Talk["Insert"]) => {
   return data;
 };
 
+const deleteChat = async (talkId: number) => {
+  const { data } = await supabase
+    .from("chat")
+    .select("chatId")
+    .eq("talkId", talkId);
+
+  if (data) {
+    await Promise.all(
+      data.map(
+        async (item) =>
+          await supabase.from("notification").delete().eq("chatId", item.chatId)
+      )
+    );
+    await supabase.from("chat").delete().eq("talkId", talkId);
+  }
+};
 const deleteTalk = async (talkId: number) => {
-  await supabase.from("chat").delete().eq("talkId", talkId);
+  await deleteChat(talkId);
   const { data, error } = await supabase
     .from("talk")
     .delete()
